:truck: Move package to 2021 archive: `utils`
