+++ pytudes/_2021/utils/linked_list.py
