+++ pytudes/educative/grokking_the_coding_interview/fast_and_slow_pointers/_1__linked_list_cycle__easy.py
@@ -10,7 +10,11 @@ See Also:
 
 """
 
-from pytudes.utils.linked_list import ListNode, NodeType, convert_list_to_linked_list
+from pytudes._2021.utils.linked_list import (
+    ListNode,
+    NodeType,
+    convert_list_to_linked_list,
+)
 
 
 def has_cycle(head: NodeType) -> bool:
