+++ pytudes/leetcode/blind_75/linked_list/_2106__reverse_linked_list__easy.py
@@ -7,7 +7,7 @@ See Also:
     - pytudes/educative/grokking_the_coding_interview/in_place_reversal_of_a_linked_list/_1__reverse_a_linked_list__easy.py
 
 """
-from pytudes.utils.linked_list import ListNode, convert_list_to_linked_list
+from pytudes._2021.utils.linked_list import ListNode, convert_list_to_linked_list
 
 ListNode = ListNode
 # Definition for singly-linked list.
