+++ pytudes/leetcode/blind_75/tree/_102__binary_tree_level_order_traversal__medium.py
@@ -11,7 +11,7 @@ See Also:
 from collections import deque
 from typing import Optional
 
-from pytudes.utils import binary_tree
+from pytudes._2021.utils import binary_tree
 
 # Definition for a binary tree node.
 # class TreeNode:
