+++ pytudes/leetcode/blind_75/tree/_104__maximum_depth_of_binary_tree__easy.py
@@ -8,7 +8,7 @@ Examples:
 
 from typing import Optional
 
-from pytudes.utils import binary_tree
+from pytudes._2021.utils import binary_tree
 
 # Definition for a binary tree node.
 # class TreeNode:
