+++ pytudes/leetcode/leetcode_patterns/fast_and_slow_pointers/_141__linked_list_cycle__easy.py
@@ -10,7 +10,7 @@ Examples:
     False
 
 """
-from pytudes.utils import linked_list
+from pytudes._2021.utils import linked_list
 
 ListNode = linked_list.ListNode
 # Definition for singly-linked list.
