+++ pytudes/leetcode/leetcode_patterns/fast_and_slow_pointers/_876__middle_of_the_linked_list__easy.py
@@ -5,7 +5,7 @@ Examples:
 
 """
 
-from pytudes.utils import linked_list
+from pytudes._2021.utils import linked_list
 
 ListNode = linked_list.ListNode
