+++ pytudes/leetcode/leetcode_patterns/in_place_reversal_of_a_linked_list/_206__reverse_linked_list__easy.py
@@ -9,7 +9,7 @@ Examples:
 
 """
 
-from pytudes.utils import linked_list
+from pytudes._2021.utils import linked_list
 
 ListNode = linked_list.ListNode
